# faster rand generation for Chisq with 1, 3 and 4 df